# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API сервиса сокращения URL",
        "title": "Go Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "post": {
                "description": "Создает короткую версию переданного URL",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Сократить URL",
                "parameters": [
                    {
                        "description": "Оригинальный URL для сокращения",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Сокращенный URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "URL уже был сокращен ранее",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/internal/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает количество URL и пользователей в сервисе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Получить статистику",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP адрес клиента",
                        "name": "X-Real-IP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Stats"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "description": "Создает короткую версию переданного URL (JSON формат)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Сократить URL (JSON)",
                "parameters": [
                    {
                        "description": "Запрос на сокращение URL",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestShortenAPI"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseShortenAPI"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseShortenAPI"
                        }
                    }
                }
            }
        },
        "/api/shorten/batch": {
            "post": {
                "description": "Создает короткие версии для списка URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Пакетное сокращение URL",
                "parameters": [
                    {
                        "description": "Список URL для сокращения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RequestShortenAPIBatch"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseShortenAPIBatch"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/urls": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает все сокращенные URL, созданные текущим пользователем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Получить URL пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResponseShortenAPIUser"
                            }
                        }
                    },
                    "204": {
                        "description": "Нет сохраненных URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Помечает указанные URL как удаленные (асинхронная операция)",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Удалить URL пользователя",
                "parameters": [
                    {
                        "description": "Список ID URL для удаления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Запрос на удаление принят",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Проверяет соединение с базой данных",
                "tags": [
                    "Storage"
                ],
                "summary": "Проверка состояния",
                "responses": {
                    "200": {
                        "description": "База данных доступна",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "База данных недоступна",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "Перенаправляет на оригинальный URL по его сокращенному ID",
                "tags": [
                    "URL"
                ],
                "summary": "Получить оригинальный URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID сокращенного URL",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Перенаправление на оригинальный URL"
                    },
                    "404": {
                        "description": "URL не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "URL был удален",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.RequestShortenAPI": {
            "description": "Запрос на создание сокращенного URL",
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://mail.ru/"
                }
            }
        },
        "models.RequestShortenAPIBatch": {
            "description": "Элемент пакетного запроса на сокращение URL",
            "type": "object",
            "properties": {
                "correlation_id": {
                    "type": "string",
                    "example": "eefbcef4-3940-5a38-b2f0-877152a6d471"
                },
                "original_url": {
                    "type": "string",
                    "example": "https://mail.ru/"
                }
            }
        },
        "models.ResponseShortenAPI": {
            "description": "Сокращенный URL",
            "type": "object",
            "properties": {
                "result": {
                    "type": "string",
                    "example": "https://localhost:8080/614904e8-28e8-5fec-aef9-56c6713e3107"
                }
            }
        },
        "models.ResponseShortenAPIBatch": {
            "description": "Элемент пакетного ответа с сокращенным URL",
            "type": "object",
            "properties": {
                "correlation_id": {
                    "type": "string",
                    "example": "eefbcef4-3940-5a38-b2f0-877152a6d471"
                },
                "short_url": {
                    "type": "string",
                    "example": "https://localhost:8080/614904e8-28e8-5fec-aef9-56c6713e3107"
                }
            }
        },
        "models.ResponseShortenAPIUser": {
            "description": "Информация о сокращенном URL пользователя",
            "type": "object",
            "properties": {
                "original_url": {
                    "type": "string",
                    "example": "https://mail.ru/"
                },
                "short_url": {
                    "type": "string",
                    "example": "https://localhost:8080/614904e8-28e8-5fec-aef9-56c6713e3107"
                }
            }
        },
        "models.Stats": {
            "description": "Информация о кол-ве пользователей в системе и сокращенных URL",
            "type": "object",
            "properties": {
                "urls": {
                    "type": "integer",
                    "example": 10
                },
                "users": {
                    "type": "integer",
                    "example": 5
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}